requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib
bcrypt
sqlalchemy
aiosqlite
langchain
langchain_community
langgraph
langchain-ollama
langchain-milvus
pymilvus
python-dotenv
pydantic[email]
pydantic-settings
numpy
pandas
requests
argon2_cffi